Refactor: Apply DIP to CepRepository using interface and injection
- Turns CepRepository into an interface and creates a concrete implementation DatabaseCepRepository, allowing dependency injection and decoupling the service from the repository.
